Move focused debug synthesis into pod-stdb
