+++ crates/pod-net/src/client_stdb.rs
@@ -39,7 +39,7 @@
 //! }
 //! ```
 
-use std::collections::{HashMap, HashSet};
+use std::collections::HashSet;
 use std::fmt;
 
 use glam::Vec2;
@@ -48,10 +48,6 @@ use uuid::Uuid;
 use pod_core::action::{AbilityTarget, Action, SpeakVolume as CoreSpeakVolume};
 use pod_core::event::{Event, GameEvent};
 use pod_core::id::{AgentId, EntityId};
-use pod_core::{
-    decode_toon_document, AgentTickRollup, AgentToolCallEvent, FocusedEntityDebugSummary,
-    ToolCallStatus,
-};
 
 use pod_stdb::client::{
     CachedEntity, StdbClient, StdbClientConfig, StdbConnectionMode, StdbError, StdbEvent,
@@ -411,7 +407,6 @@ pub struct SpacetimeDBClient {
     local_snapshot: Option<WorldSnapshot>,
     last_debug_telemetry_json: Option<String>,
     pending_debug_documents: Vec<String>,
-    focused_debug_summaries: HashMap<u64, FocusedEntityDebugSummary>,
     render_buffer: SnapshotInterpolationBuffer,
     render_clock: RenderClock,
     welcome_sent: bool,
@@ -430,7 +425,6 @@ impl SpacetimeDBClient {
             local_snapshot: None,
             last_debug_telemetry_json: None,
             pending_debug_documents: Vec::new(),
-            focused_debug_summaries: HashMap::new(),
             render_buffer: SnapshotInterpolationBuffer::default(),
             render_clock: RenderClock::default(),
             welcome_sent: false,
@@ -727,30 +721,18 @@ impl SpacetimeDBClient {
                 StdbEvent::AgentToolCallEventReceived { document, .. } => {
                     self.pending_debug_documents.push(document.clone());
                     messages.push(ServerMessage::DebugDocument {
-                        document: document.clone(),
+                        document,
                     });
-                    if let Some(summary_document) =
-                        self.update_focused_summary_from_tool_call_document(&document)
-                    {
-                        self.pending_debug_documents.push(summary_document.clone());
-                        messages.push(ServerMessage::DebugDocument {
-                            document: summary_document,
-                        });
-                    }
                 }
                 StdbEvent::AgentTickRollupReceived { document, .. } => {
                     self.pending_debug_documents.push(document.clone());
                     messages.push(ServerMessage::DebugDocument {
-                        document: document.clone(),
+                        document,
                     });
-                    if let Some(summary_document) =
-                        self.update_focused_summary_from_rollup_document(&document)
-                    {
-                        self.pending_debug_documents.push(summary_document.clone());
-                        messages.push(ServerMessage::DebugDocument {
-                            document: summary_document,
-                        });
-                    }
+                }
+                StdbEvent::FocusedEntityDebugSummaryReceived { document, .. } => {
+                    self.pending_debug_documents.push(document.clone());
+                    messages.push(ServerMessage::DebugDocument { document });
                 }
 
                 // ── Tick advancement ──
@@ -788,7 +770,6 @@ impl SpacetimeDBClient {
         self.last_emitted_tick = 0;
         self.last_debug_telemetry_json = None;
         self.pending_debug_documents.clear();
-        self.focused_debug_summaries.clear();
         self.subscriptions.reset_connection();
         self.clear_presentation_state();
     }
@@ -975,93 +956,6 @@ impl SpacetimeDBClient {
     /// same browser/editor consumer path as tick telemetry.
     pub fn push_debug_document(&mut self, document: String) {
         self.pending_debug_documents.push(document);
-    }
-
-    fn focused_debug_summary_entry(&mut self, entity_id: u64) -> &mut FocusedEntityDebugSummary {
-        let shard_id = self.inner.config().db_name.clone();
-        self.focused_debug_summaries
-            .entry(entity_id)
-            .or_insert_with(|| FocusedEntityDebugSummary {
-                shard_id,
-                entity_id,
-                ..Default::default()
-            })
-    }
-
-    fn refresh_focused_summary_notes(summary: &mut FocusedEntityDebugSummary) {
-        summary.notes.clear();
-        if summary.tool_error_count > 0 {
-            summary
-                .notes
-                .push(format!("{} tool-call errors retained", summary.tool_error_count));
-        }
-        if summary.rejected_action_count > 0 {
-            summary.notes.push(format!(
-                "{} rejected actions retained",
-                summary.rejected_action_count
-            ));
-        }
-    }
-
-    fn update_focused_summary_from_tool_call_document(&mut self, document: &str) -> Option<String> {
-        let event: AgentToolCallEvent =
-            match decode_toon_document(document, "agent_tool_call_event") {
-                Ok(event) => event,
-                Err(error) => {
-                    log::warn!(
-                        "[pod-net stdb] Failed to decode agent_tool_call_event for focused summary synthesis: {error}"
-                    );
-                    return None;
-                }
-            };
-
-        let summary = self.focused_debug_summary_entry(event.agent_entity_id);
-        let prior_tool_count = summary.tool_call_count;
-        summary.latest_tick = summary.latest_tick.max(event.tick);
-        summary.tool_call_count += 1;
-        summary.average_tool_latency_ms = if prior_tool_count == 0 {
-            event.trace.latency_ms as f32
-        } else {
-            ((summary.average_tool_latency_ms * prior_tool_count as f32)
-                + event.trace.latency_ms as f32)
-                / summary.tool_call_count as f32
-        };
-        summary.latest_tool_name = Some(event.trace.tool_name.clone());
-        summary.latest_tool_status = Some(format!("{:?}", event.trace.status));
-        summary.latest_tool_error = event.trace.error_message.clone();
-        if !matches!(
-            event.trace.status,
-            ToolCallStatus::Requested | ToolCallStatus::Succeeded
-        ) {
-            summary.tool_error_count += 1;
-        }
-        Self::refresh_focused_summary_notes(summary);
-        Some(summary.to_toon_document())
-    }
-
-    fn update_focused_summary_from_rollup_document(&mut self, document: &str) -> Option<String> {
-        let rollup: AgentTickRollup = match decode_toon_document(document, "agent_tick_rollup") {
-            Ok(rollup) => rollup,
-            Err(error) => {
-                log::warn!(
-                    "[pod-net stdb] Failed to decode agent_tick_rollup for focused summary synthesis: {error}"
-                );
-                return None;
-            }
-        };
-
-        let summary = self.focused_debug_summary_entry(rollup.agent_entity_id);
-        summary.latest_tick = summary.latest_tick.max(rollup.tick_end);
-        summary.tool_call_count = rollup.tool_call_count as usize;
-        summary.tool_error_count = rollup.tool_error_count as usize;
-        summary.rejected_action_count = rollup.rejected_action_count as usize;
-        summary.total_distance = rollup.total_distance;
-        summary.average_tool_latency_ms = rollup.average_tool_latency_ms;
-        summary.visible_entity_count = rollup.visible_entity_count as usize;
-        summary.audible_event_count = rollup.audible_event_count as usize;
-        summary.message_count = rollup.message_count as usize;
-        Self::refresh_focused_summary_notes(summary);
-        Some(summary.to_toon_document())
     }
 
     /// Inspect the local rollback/replay path from a chosen retained tick.
