+++ progress.md
@@ -35,3 +35,4 @@ Original prompt: its the graphics and the worlds scene, everything is piled up o
 - Moved focused-entity debug aggregation into shared `pod-core` runtime helpers and reused that path from both shard ops tooling and direct-connect runtime code, removing drift between server-side summary producers.
 - Added `SetDebugFocus { entity_id }` to `pod-net` and wired `pod-web` + `pod-editor` to consume `focused_entity_debug_summary` TOON docs, so browser/editor selections now line up with shard-delivered focused debug state instead of only reconstructing it client-side.
 - Added focused-summary synthesis to the `pod-net` SpacetimeDB adapter path, so entity-scoped tool-call and rollup TOON streams now emit `focused_entity_debug_summary` documents with retained latest-tool metadata and rollup metrics for browser/editor parity.
+- Moved focused-summary synthesis into `pod-stdb::client` itself and reduced `pod-net` back to a thin forwarder for `FocusedEntityDebugSummaryReceived`, so shard/debug consumers share one focused-summary event source instead of adapter-local rebuilding.
